requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.1.4
numpy==1.26.2
